bugfix: remove accents for contents and structures
