+++ client/src/lib/search.ts
@@ -13,19 +13,3 @@ export const getSuggestionValue = (suggestion: SimplifiedUser | SimplifiedStruct
     (suggestion.email || "");
 }
 // : suggestion.titreMarque || suggestion.titreInformatif //+ (suggestion.titreMarque && suggestion.titreInformatif ? " - " : "") + suggestion.titreInformatif;
-
-
-export const removeAccents = (str = "") => {
-  var accents =
-    "ÀÁÂÃÄÅàáâãäåÒÓÔÕÕÖØòóôõöøÈÉÊËèéêëðÇçÐÌÍÎÏìíîïÙÚÛÜùúûüÑñŠšŸÿýŽž";
-  var accentsOut =
-    "AAAAAAaaaaaaOOOOOOOooooooEEEEeeeeeCcDIIIIiiiiUUUUuuuuNnSsYyyZz";
-  const chars = str.split("");
-  let i, x;
-  for (i = 0; i < chars.length; i++) {
-    if ((x = accents.indexOf(chars[i])) !== -1) {
-      chars[i] = accentsOut[x];
-    }
-  }
-  return chars.join("");
-};
